Popover and Stage scroll with fixed Element
Added the possibility to add the property elementPosition to a Driver.Step. When this property is set to 'fixed', the popup and the stage will be positioned 'fixed'. This fixes in issue, in which the tooltip would not stay on the highlighted Element, when the user scrolled.

It was not enough to set the position to 'fixed' on all 'fixed' highlighted elements, because a child of 'fixed' position element is itself not 'fixed' but behaves that way because of its parent.